Move creators to structures
