+++ packages/backend-core/tests/utilities/structures/index.ts
@@ -9,3 +9,4 @@ export * as sso from "./sso"
 export * as tenant from "./tenants"
 export * as users from "./users"
 export { generator } from "./generator"
+export * as scim from "./scim"
